requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
mongoengine==0.29.3
PyJWT==2.7.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.2.2
pydantic[email]==2.13.0
httpx==0.27.0